Move subnormal tests into seperate file
